fixing bugs and refactoring

+++ include/LSH.h
@@ -19,7 +19,7 @@ private:
     uint32_t calculate_g(const std::vector<uint8_t> &x, const std::vector<std::vector<int>> &S);
 
 public:
-    LSH(int k, int L, Data &data, uint32_t w = 10, uint32_t m = uint32_t(pow(2, 30) - 5));
+    LSH(int k, int L, Data &data, uint32_t w = 10, uint32_t m = pow(2, 30) - 5);
     ~LSH();
 
     int Run(const std::vector<uint8_t> &query, std::ofstream &outputFile, const int &N);
